chore: refactor react dev tools to utils
